utils/SerialOTAUpdater/requirements.txt
pyserial
crc8
pymodbus
pytest-snapshot=0.9.0